incoem:new version

+++ lib/features/income/presentation/bindings/income_binding.dart
@@ -1,6 +1,5 @@
 import 'package:dio/dio.dart';
 import 'package:get/get.dart';
-import 'package:spendwise/features/auth/domain/usecases/get_user_id_usecase.dart';
 import 'package:spendwise/features/income/data/datasources/income_local_datasource.dart';
 import 'package:spendwise/features/income/data/datasources/income_local_datasources_impl.dart';
 import 'package:spendwise/features/income/data/datasources/income_remote_datasource.dart';
@@ -9,6 +8,7 @@ import 'package:spendwise/features/income/data/repositories/income_repository.da
 import 'package:spendwise/features/income/data/repositories/income_repository_impl.dart';
 import 'package:spendwise/features/income/domain/usecases/add_income_usecase.dart';
 import 'package:spendwise/features/income/domain/usecases/delete_income_usecase.dart';
+import 'package:spendwise/features/income/domain/usecases/get_all_local_incomes_usecase.dart';
 import 'package:spendwise/features/income/domain/usecases/get_incomes_usecase.dart';
 import 'package:spendwise/features/income/domain/usecases/update_income_usecase.dart';
 import 'package:spendwise/features/income/presentation/manager/income_controller.dart';
@@ -20,32 +20,52 @@ import 'package:spendwise/features/wallet/presentation/manager/wallet_controller
 class IncomeBinding extends Bindings {
   @override
   void dependencies() {
-    Get.lazyPut<IncomeRemoteDatasource>(
-      () => IncomeRemoteDatasourceImpl(dio: Get.find<Dio>()),
-    );
-    Get.lazyPut<IncomeLocalDataSource>(() {
-      final datasource = IncomeLocalDataSourceImpl();
-      return datasource;
-    });
-    Get.lazyPut<IncomeRepository>(
-      () => IncomeRepositoryImpl(
-        localDataSource: Get.find<IncomeLocalDataSource>(),
-        remoteDatasource: Get.find<IncomeRemoteDatasource>(),
-      ),
-    );
-    Get.lazyPut(() => AddIncomeUsecase(Get.find<IncomeRepository>()));
-    Get.lazyPut(() => GetIncomesUsecase(Get.find<IncomeRepository>()));
-    Get.lazyPut(() => UpdateIncomeUseCase(Get.find<IncomeRepository>()));
-    Get.lazyPut(() => DeleteIncomeUseCase(Get.find<IncomeRepository>()));
-    Get.put(
-      IncomeController(
-        addIncomeUseCase: Get.find<AddIncomeUsecase>(),
-        getIncomesUseCase: Get.find<GetIncomesUsecase>(),
-        walletController: Get.find<WalletController>(),
-        tagController: Get.find<TagController>(),
-        updateIncomeUseCase: Get.find<UpdateIncomeUseCase>(),
-        deleteIncomeUseCase: Get.find<DeleteIncomeUseCase>(),
-      ),
-    );
+    if (!Get.isRegistered<IncomeRemoteDatasource>()) {
+      Get.lazyPut<IncomeRemoteDatasource>(
+        () => IncomeRemoteDatasourceImpl(dio: Get.find<Dio>()),
+      );
+    }
+    if (!Get.isRegistered<IncomeLocalDataSource>()) {
+      Get.lazyPut<IncomeLocalDataSource>(() => IncomeLocalDataSourceImpl());
+    }
+    if (!Get.isRegistered<IncomeRepository>()) {
+      Get.lazyPut<IncomeRepository>(
+        () => IncomeRepositoryImpl(
+          localDataSource: Get.find<IncomeLocalDataSource>(),
+          remoteDatasource: Get.find<IncomeRemoteDatasource>(),
+        ),
+      );
+    }
+    if (!Get.isRegistered<AddIncomeUsecase>()) {
+      Get.lazyPut(() => AddIncomeUsecase(Get.find<IncomeRepository>()));
+    }
+    if (!Get.isRegistered<GetIncomesUsecase>()) {
+      Get.lazyPut(() => GetIncomesUsecase(Get.find<IncomeRepository>()));
+    }
+    if (!Get.isRegistered<GetAllLocalIncomesUsecase>()) {
+      Get.lazyPut(
+        () => GetAllLocalIncomesUsecase(Get.find<IncomeRepository>()),
+      );
+    }
+    if (!Get.isRegistered<UpdateIncomeUseCase>()) {
+      Get.lazyPut(() => UpdateIncomeUseCase(Get.find<IncomeRepository>()));
+    }
+    if (!Get.isRegistered<DeleteIncomeUseCase>()) {
+      Get.lazyPut(() => DeleteIncomeUseCase(Get.find<IncomeRepository>()));
+    }
+    if (!Get.isRegistered<IncomeController>()) {
+      Get.lazyPut(
+        () => IncomeController(
+          addIncomeUseCase: Get.find<AddIncomeUsecase>(),
+          getIncomesUseCase: Get.find<GetIncomesUsecase>(),
+          getAllLocalIncomesUsecase: Get.find<GetAllLocalIncomesUsecase>(),
+          walletController: Get.find<WalletController>(),
+          tagController: Get.find<TagController>(),
+          updateIncomeUseCase: Get.find<UpdateIncomeUseCase>(),
+          deleteIncomeUseCase: Get.find<DeleteIncomeUseCase>(),
+        ),
+        fenix: true,
+      );
+    }
   }
 }
